service/requirements.txt
dnspython==2.4.2
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pymongo==4.6.1
redis==5.0.1
